[DTensor] fix F.one_hot
F.one_hot(dtensor) used to run into a mixed DTensor-Tensor operation due
to an arange call creating a new Tensor (not DTensor). This PR fixes it
by allowing implicit replication of Tensors for the arange call and the
one consumer of the arange call (the at::eq call).

Test Plan:
- new test. Also, F.one_hot(num_classes=-1) is broken so we skip that.

ghstack-source-id: 613b6362749eaccf2c0e4d03004002ba9d10fff0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/162307