lint: Only include files in pytorch
We were seeing instances of stdlib files in clang-tidy output so this
just essentially removes them from the things that lintrunner will
report up. Longer term fix here would be to just modify the clang-tidy
configuration in order to do the correct thing here but that requires a
bit more investigation as to why this is only happening in CI and is not
reproduceable locally.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: a1a34dd8cd769b85618a1fac578a6cb5986f7f2d
Pull-Request: https://github.com/pytorch/pytorch/pull/164008