ci: Removing ROCm tests from trunk.
Had a conversation with the AMD team today and I think we are all in
agreement that the current state of queueing for AMD is beyond where
we'd like to be for there to be blocking CI for ROCm.

Moving the representative testing jobs for this into the ciflow/rocm
workflow.

I'd love for these to be back in trunk if we can get to a state where
our queueing metrics are below an hour for ROCm infrastructure.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: a8acca9b0be287399505cd5555ee481981a15a95
Pull-Request: https://github.com/pytorch/pytorch/pull/164585
Signed-off-by: Eli Uriegas <eliuriegas@meta.com>