Use PyObject_GetOptionalAttrString in PyObject_FastGetAttrString when available
Python 3.13 added PyObject_GetOptionalAttrString. I'm not 100% certain that it is strictly better than the old approach in all cases, but based on documentation/comments it seems to be meant for this type of use, and it's faster when I profile torchtitan training (which gets to checking for the __torch_function__ attr on some object frequently enough to notice, but wastes a bunch of time generating exceptions that we then suppressed here).

ghstack-source-id: 2215645d4477a48aefa5ba5dd5792f5d482274fb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/164624