[export][dynamo] Fallback to slowpath for MultiHeadAttention for compile
In #106824, export decided to slow-path for MultiHeadAttention module (look into the PR description as to why). But that PR eventually caused a divergence between Dynamo and export.

Today, strict-export does not inline into builtin modules (like MultiHeadAttention), and therefore make_fx sees the original nn.Module and takes the slow path. But compile inlines into the nn module, and at this time the condition _is_make_fx_tracing is False. As a result, Dynamo takes a fast path, resulting in a different op being called.

This divergence is undesirable. There are 2 ways to fix it

Make export take the fast path - As explained in the [pytorch] Disable fast path in MultiheadAttention in Export #106824 , this might be difficult. So, we go to (2)
Make compile as well take the slow path - This is easy to implement. The con here is that Pytorch eager and compile will use different operators, which can cause numerics issues etc.
Since (2) is easy to do, we will follow this path. We are tracking the issue in #164062

ghstack-source-id: 3130df320b3d82309dbd999ee307092354b3e65b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/164721