Used a fixed-length array to represent the interpreter stack
CPython represents the stack of a frame as a fixed length C array + a stack pointer. The pointer points to the next free slot. When the interpreter executes a PUSH or a POP operation, it increments or decrements the pointer, without having to resize the underlying array. This allows fast O(1) append/pop operations.

ghstack-source-id: 8447538b3b31d3cc74c6f385d2a9f609fc690b7e
Pull-Request: https://github.com/pytorch/pytorch/pull/162724