Used a fixed-length array to represent the interpreter stack
CPython represents the stack of a frame as a fixed length C array + a stack pointer. The pointer points to the next free slot. When the interpreter executes a PUSH or a POP operation, it increments or decrements the pointer, without having to resize the underlying array. This allows fast O(1) append/pop operations.


ghstack-source-id: 396842d4d5b53f2bcc3ed0f65c6c989b80f8b07e
Pull-Request: https://github.com/pytorch/pytorch/pull/165587