Used a fixed-length array to represent the interpreter stack
CPython represents the stack of a frame as a fixed length C array + a stack pointer. The pointer points to the next free slot. When the interpreter executes a PUSH or a POP operation, it increments or decrements the pointer, without having to resize the underlying array. This allows fast O(1) append/pop operations.

ghstack-source-id: e1d69e4ccef58c0b5f1a11d5505c982cd5c99b02
Pull-Request: https://github.com/pytorch/pytorch/pull/162724